--- cs2030s/.idea/$PROJECT_FILE$
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="AnalysisProjectProfileManager">
-    <option name="PROJECT_PROFILE" />
-    <option name="USE_PROJECT_LEVEL_SETTINGS" value="false" />
-    <list size="0" />
-  </component>
-  <component name="SuppressionsComponent">
-    <option name="suppComments" value="[]" />
-  </component>
-</project>
