--- cs2030s/.idea/qaplug_profiles.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="AnalysisProjectProfileManager">
-    <option name="PROJECT_PROFILE" value="Project Default" />
-    <option name="USE_PROJECT_LEVEL_SETTINGS" value="false" />
-    <scopes />
-    <profiles>
-      <profile profile_name="Project Default" version="1.0" is_locked="false" />
-    </profiles>
-    <list size="0" />
-  </component>
-</project>
