--- cs2030s/src/Main.java
@@ -1,6 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-        Array<String> s = new Array<String>(10);
-    }
-}
